Move N+1 API Calls detector to own file
