+++ src/sentry/utils/performance_issues/detectors/__init__.py
@@ -1,0 +1,1 @@
+from .n_plus_one_api_calls_detector import NPlusOneAPICallsDetector  # NOQA
